+++ packages/core/src/PartialExecutionUtils/__tests__/findStartNodes.test.ts
@@ -1,5 +1,6 @@
 // NOTE: Diagrams in this file have been created with https://asciiflow.com/#/
-// If you update the tests please update the diagrams as well.
+// If you update the tests, please update the diagrams as well.
+// If you add a test, please create a new diagram.
 //
 // Map
 // 0  means the output has no run data
@@ -8,135 +9,10 @@
 // XX denotes that the node is disabled
 // PD denotes that the node has pinned data
 
-import type { IConnections, INode, IPinData, IRunData } from 'n8n-workflow';
-import { NodeConnectionType } from 'n8n-workflow';
-import { DirectedGraph, findStartNodes, findSubgraph, isDirty } from '../utils';
-import { createNodeData, toITaskData, defaultWorkflowParameter } from './helpers';
-
-test('toITaskData', function () {
-	expect(toITaskData([{ data: { value: 1 } }])).toEqual({
-		executionStatus: 'success',
-		executionTime: 0,
-		source: [],
-		startTime: 0,
-		data: {
-			main: [[{ json: { value: 1 } }]],
-		},
-	});
-
-	expect(toITaskData([{ data: { value: 1 }, outputIndex: 1 }])).toEqual({
-		executionStatus: 'success',
-		executionTime: 0,
-		source: [],
-		startTime: 0,
-		data: {
-			main: [null, [{ json: { value: 1 } }]],
-		},
-	});
-
-	expect(
-		toITaskData([
-			{ data: { value: 1 }, outputIndex: 1, nodeConnectionType: NodeConnectionType.AiAgent },
-		]),
-	).toEqual({
-		executionStatus: 'success',
-		executionTime: 0,
-		source: [],
-		startTime: 0,
-		data: {
-			[NodeConnectionType.AiAgent]: [null, [{ json: { value: 1 } }]],
-		},
-	});
-
-	expect(
-		toITaskData([
-			{ data: { value: 1 }, outputIndex: 0 },
-			{ data: { value: 2 }, outputIndex: 1 },
-		]),
-	).toEqual({
-		executionStatus: 'success',
-		executionTime: 0,
-		startTime: 0,
-		source: [],
-		data: {
-			main: [
-				[
-					{
-						json: { value: 1 },
-					},
-				],
-				[
-					{
-						json: { value: 2 },
-					},
-				],
-			],
-		},
-	});
-});
-
-type Connection = {
-	from: INode;
-	to: INode;
-	type?: NodeConnectionType;
-	outputIndex?: number;
-	inputIndex?: number;
-};
-
-function toIConnections(connections: Connection[]): IConnections {
-	const result: IConnections = {};
-
-	for (const connection of connections) {
-		const type = connection.type ?? NodeConnectionType.Main;
-		const outputIndex = connection.outputIndex ?? 0;
-		const inputIndex = connection.inputIndex ?? 0;
-
-		result[connection.from.name] = result[connection.from.name] ?? {
-			[type]: [],
-		};
-		const resultConnection = result[connection.from.name];
-		resultConnection[type][outputIndex] = resultConnection[type][outputIndex] ?? [];
-		const group = resultConnection[type][outputIndex];
-
-		group.push({
-			node: connection.to.name,
-			type,
-			index: inputIndex,
-		});
-	}
-
-	return result;
-}
-
-test('toIConnections', () => {
-	const node1 = createNodeData({ name: 'Basic Node 1' });
-	const node2 = createNodeData({ name: 'Basic Node 2' });
-
-	expect(
-		toIConnections([{ from: node1, to: node2, type: NodeConnectionType.Main, outputIndex: 0 }]),
-	).toEqual({
-		[node1.name]: {
-			// output group
-			main: [
-				// first output
-				[
-					// first connection
-					{
-						node: node2.name,
-						type: NodeConnectionType.Main,
-						index: 0,
-					},
-				],
-			],
-		},
-	});
-});
-
-//export interface INodeTypes {
-//	getByName(nodeType: string): INodeType | IVersionedNodeType;
-//	getByNameAndVersion(nodeType: string, version?: number): INodeType;
-//	getKnownTypes(): IDataObject;
-//}
+import type { IPinData, IRunData } from 'n8n-workflow';
+import { createNodeData, toITaskData } from './helpers';
+import { findStartNodes, isDirty } from '../findStartNodes';
+import { DirectedGraph } from '../DirectedGraph';
 
 describe('isDirty', () => {
 	test("if the node has pinned data it's not dirty", () => {
@@ -562,310 +438,6 @@ describe('findStartNodes', () => {
 				previousNodeRun: 0,
 				previousNodeOutput: 0,
 			},
-		});
-	});
-});
-
-describe('findSubgraph2', () => {
-	//                 ►►
-	//  ┌───────┐     ┌───────────┐
-	//  │trigger├────►│destination│
-	//  └───────┘     └───────────┘
-	test('simple', () => {
-		const trigger = createNodeData({ name: 'trigger' });
-		const destination = createNodeData({ name: 'destination' });
-
-		const graph = new DirectedGraph()
-			.addNodes(trigger, destination)
-			.addConnections({ from: trigger, to: destination });
-
-		const subgraph = findSubgraph(graph, destination, trigger);
-
-		expect(subgraph).toEqual(graph);
-	});
-
-	//                     ►►
-	//  ┌───────┐         ┌───────────┐
-	//  │       ├────────►│           │
-	//  │trigger│         │destination│
-	//  │       ├────────►│           │
-	//  └───────┘         └───────────┘
-	test('multiple connections', () => {
-		const ifNode = createNodeData({ name: 'If' });
-		const noOp = createNodeData({ name: 'noOp' });
-
-		const graph = new DirectedGraph()
-			.addNodes(ifNode, noOp)
-			.addConnections(
-				{ from: ifNode, to: noOp, outputIndex: 0 },
-				{ from: ifNode, to: noOp, outputIndex: 1 },
-			);
-
-		const subgraph = findSubgraph(graph, noOp, ifNode);
-
-		expect(subgraph).toEqual(graph);
-	});
-
-	//                     ►►
-	//  ┌───────┐         ┌───────────┐
-	//  │       ├────────►│           │      ┌────┐
-	//  │trigger│         │destination├─────►│node│
-	//  │       ├────────►│           │      └────┘
-	//  └───────┘         └───────────┘
-	test('disregard nodes after destination', () => {
-		const trigger = createNodeData({ name: 'trigger' });
-		const destination = createNodeData({ name: 'destination' });
-		const node = createNodeData({ name: 'node' });
-
-		const graph = new DirectedGraph()
-			.addNodes(trigger, destination, node)
-			.addConnections({ from: trigger, to: destination }, { from: destination, to: node });
-
-		const subgraph = findSubgraph(graph, destination, trigger);
-
-		expect(subgraph).toEqual(
-			new DirectedGraph()
-				.addNodes(trigger, destination)
-				.addConnections({ from: trigger, to: destination }),
-		);
-	});
-
-	//                     XX
-	//  ┌───────┐         ┌────────┐       ►►
-	//  │       ├────────►│        │      ┌───────────┐
-	//  │trigger│         │disabled├─────►│destination│
-	//  │       ├────────►│        │      └───────────┘
-	//  └───────┘         └────────┘
-	test('skip disabled nodes', () => {
-		const trigger = createNodeData({ name: 'trigger' });
-		const disabled = createNodeData({ name: 'disabled', disabled: true });
-		const destination = createNodeData({ name: 'destination' });
-
-		const graph = new DirectedGraph()
-			.addNodes(trigger, disabled, destination)
-			.addConnections({ from: trigger, to: disabled }, { from: disabled, to: destination });
-
-		const subgraph = findSubgraph(graph, destination, trigger);
-
-		expect(subgraph).toEqual(
-			new DirectedGraph()
-				.addNodes(trigger, destination)
-				.addConnections({ from: trigger, to: destination }),
-		);
-	});
-
-	//                                ►►
-	//  ┌───────┐       ┌─────┐      ┌─────┐
-	//  │Trigger├───┬──►│Node1├───┬─►│Node2│
-	//  └───────┘   │   └─────┘   │  └─────┘
-	//              │             │
-	//              └─────────────┘
-	test('terminates when called with graph that contains cycles', () => {
-		//
-		// ARRANGE
-		//
-		const trigger = createNodeData({ name: 'trigger' });
-		const node1 = createNodeData({ name: 'node1' });
-		const node2 = createNodeData({ name: 'node2' });
-		const graph = new DirectedGraph()
-			.addNodes(trigger, node1, node2)
-			.addConnections(
-				{ from: trigger, to: node1 },
-				{ from: node1, to: node1 },
-				{ from: node1, to: node2 },
-			);
-
-		//
-		// ACT
-		//
-		const subgraph = findSubgraph(graph, node2, trigger);
-
-		//
-		// ASSERT
-		//
-		expect(subgraph).toEqual(graph);
-	});
-
-	//                ►►
-	//  ┌───────┐     ┌─────┐
-	//  │Trigger├──┬─►│Node1│
-	//  └───────┘  │  └─────┘
-	//             │
-	//  ┌─────┐    │
-	//  │Node2├────┘
-	//  └─────┘
-	test('terminates when called with graph that contains cycles', () => {
-		//
-		// ARRANGE
-		//
-		const trigger = createNodeData({ name: 'trigger' });
-		const node1 = createNodeData({ name: 'node1' });
-		const node2 = createNodeData({ name: 'node2' });
-		const graph = new DirectedGraph()
-			.addNodes(trigger, node1, node2)
-			.addConnections({ from: trigger, to: node1 }, { from: node2, to: node1 });
-
-		//
-		// ACT
-		//
-		const subgraph = findSubgraph(graph, node1, trigger);
-
-		//
-		// ASSERT
-		//
-		expect(subgraph).toEqual(
-			new DirectedGraph().addNodes(trigger, node1).addConnections({ from: trigger, to: node1 }),
-		);
-	});
-
-	//                               ►►
-	//  ┌───────┐    ┌───────────┐   ┌───────────┐
-	//  │Trigger├─┬─►│Destination├──►│AnotherNode├───┐
-	//  └───────┘ │  └───────────┘   └───────────┘   │
-	//            │                                  │
-	//            └──────────────────────────────────┘
-	test('terminates if the destination node is part of a cycle', () => {
-		//
-		// ARRANGE
-		//
-		const trigger = createNodeData({ name: 'trigger' });
-		const destination = createNodeData({ name: 'destination' });
-		const anotherNode = createNodeData({ name: 'anotherNode' });
-		const graph = new DirectedGraph()
-			.addNodes(trigger, destination, anotherNode)
-			.addConnections(
-				{ from: trigger, to: destination },
-				{ from: destination, to: anotherNode },
-				{ from: anotherNode, to: destination },
-			);
-
-		//
-		// ACT
-		//
-		const subgraph = findSubgraph(graph, destination, trigger);
-
-		//
-		// ASSERT
-		//
-		expect(subgraph).toEqual(
-			new DirectedGraph()
-				.addNodes(trigger, destination)
-				.addConnections({ from: trigger, to: destination }),
-		);
-	});
-});
-
-describe('DirectedGraph', () => {
-	//     ┌─────┐    ┌─────┐   ┌─────┐
-	//  ┌─►│node1├───►│node2├──►│node3├─┐
-	//  │  └─────┘    └─────┘   └─────┘ │
-	//  │                               │
-	//  └───────────────────────────────┘
-	test('roundtrip', () => {
-		const node1 = createNodeData({ name: 'Node1' });
-		const node2 = createNodeData({ name: 'Node2' });
-		const node3 = createNodeData({ name: 'Node3' });
-
-		const graph = new DirectedGraph()
-			.addNodes(node1, node2, node3)
-			.addConnections(
-				{ from: node1, to: node2 },
-				{ from: node2, to: node3 },
-				{ from: node3, to: node1 },
-			);
-
-		expect(DirectedGraph.fromWorkflow(graph.toWorkflow({ ...defaultWorkflowParameter }))).toEqual(
-			graph,
-		);
-	});
-
-	describe('getChildren', () => {
-		//  ┌─────┐       ┌─────┐
-		//  │node1├──────►│node2│
-		//  └─────┘       └─────┘
-		test('simple', () => {
-			const from = createNodeData({ name: 'Node1' });
-			const to = createNodeData({ name: 'Node2' });
-			const graph = new DirectedGraph().addNodes(from, to).addConnections({ from, to });
-
-			const children = graph.getChildren(from);
-			expect(children).toHaveLength(1);
-			expect(children).toContainEqual({
-				from,
-				to,
-				inputIndex: 0,
-				outputIndex: 0,
-				type: NodeConnectionType.Main,
-			});
-		});
-		//  ┌─────┐
-		//  │     ├────┐  ┌─────┐
-		//  │node1│    ├─►│node2│
-		//  │     ├────┘  └─────┘
-		//  └─────┘
-		test('medium', () => {
-			const from = createNodeData({ name: 'Node1' });
-			const to = createNodeData({ name: 'Node2' });
-			const graph = new DirectedGraph()
-				.addNodes(from, to)
-				.addConnections({ from, to, outputIndex: 0 }, { from, to, outputIndex: 1 });
-
-			const children = graph.getChildren(from);
-			expect(children).toHaveLength(2);
-			expect(children).toContainEqual({
-				from,
-				to,
-				inputIndex: 0,
-				outputIndex: 0,
-				type: NodeConnectionType.Main,
-			});
-			expect(children).toContainEqual({
-				from,
-				to,
-				inputIndex: 0,
-				outputIndex: 1,
-				type: NodeConnectionType.Main,
-			});
-		});
-
-		//     ┌─────┐       ┌─────┐
-		//  ┌─►│node1├──────►│node2├──┐
-		//  │  └─────┘       └─────┘  │
-		//  │                         │
-		//  └─────────────────────────┘
-		test('terminates if the graph has cycles', () => {
-			//
-			// ARRANGE
-			//
-			const node1 = createNodeData({ name: 'node1' });
-			const node2 = createNodeData({ name: 'node2' });
-			const graph = new DirectedGraph()
-				.addNodes(node1, node2)
-				.addConnections({ from: node1, to: node2 }, { from: node2, to: node2 });
-
-			//
-			// ACT
-			//
-			const children = graph.getChildren(node1);
-
-			//
-			// ASSERT
-			//
-			expect(children).toHaveLength(2);
-			expect(children).toContainEqual({
-				from: node1,
-				to: node2,
-				inputIndex: 0,
-				outputIndex: 0,
-				type: NodeConnectionType.Main,
-			});
-			expect(children).toContainEqual({
-				from: node2,
-				to: node2,
-				inputIndex: 0,
-				outputIndex: 0,
-				type: NodeConnectionType.Main,
-			});
 		});
 	});
 });
