+++ packages/core/src/PartialExecutionUtils/findCycles.ts
@@ -1,0 +1,6 @@
+import type { Workflow } from 'n8n-workflow';
+
+export function findCycles(_workflow: Workflow) {
+	// TODO: implement depth first search or Tarjan's Algorithm
+	return [];
+}
