+++ packages/core/src/WorkflowExecute.ts
@@ -50,15 +50,14 @@ import get from 'lodash/get';
 import * as NodeExecuteFunctions from './NodeExecuteFunctions';
 
 import * as a from 'assert';
+import { recreateNodeExecutionStack } from './PartialExecutionUtils/recreateNodeExecutionStack';
 import {
 	DirectedGraph,
 	findCycles,
 	findStartNodes,
 	findSubgraph,
 	findTriggerForPartialExecution,
-} from './utils';
-// TODO: move into it's own folder
-import { recreateNodeExecutionStack } from './utils-2';
+} from './PartialExecutionUtils';
 
 export class WorkflowExecute {
 	private status: ExecutionStatus = 'new';
